Subject: gcc format strings: Make GCC check format strings, fix found problems

+++ lib/lws-plat-unix.c
@@ -685,8 +685,8 @@ lws_plat_init(struct lws_context *context,
 		return 1;
 	}
 
-	lwsl_notice(" mem: platform fd map: %5u bytes\n",
-		    sizeof(struct lws *) * context->max_fds);
+	lwsl_notice(" mem: platform fd map: %5lu bytes\n",
+		    (unsigned long)(sizeof(struct lws *) * context->max_fds));
 	fd = open(SYSTEM_RANDOM_FILEPATH, O_RDONLY);
 
 	context->fd_random = fd;
